# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Golem Workers Specification",
    "version": "0.5.0"
  },
  "paths": {
...
    "/get-proposals": {
      "post": {
        "tags": [
          "misc"
        ],
        "summary": "Get Proposals",
        "description": "Get proposals based on the provided parameters.",
        "operationId": "get_proposals_get_proposals_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetProposalsRequest"
              },
              "examples": {
                "minimal_cpu": {
                  "summary": "Minimal CPU",
                  "description": "This example shows how to select providers for Virtual Machine.",
                  "value": {
                    "market_config": {
                      "demand": {
                        "payloads": [
                          "golem_workers.payloads.ClusterNodePayload"
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProposalsResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Unhandled server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPGenericError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cluster": {
      "get": {
        "tags": [
          "clusters"
        ],
        "summary": "List Clusters",
        "description": "Lists available clusters",
        "operationId": "list_clusters_cluster_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "examples": [
                    [
                      "cluster1",
                      "cluster2"
                    ]
                  ],
                  "title": "Response List Clusters Cluster Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "clusters"
        ],
        "summary": "Create Cluster",
        "description": "Creates cluster and schedules its start.",
        "operationId": "create_cluster_cluster_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClusterRequest"
              },
              "examples": {
                "testnet_linear_budget_vpn_reputation": {
                  "summary": "Average usage budget, VPN and reputation (testnet)",
                  "description": "This example shows how to create a testnet cluster that support average usage budget, simple VPN network and Golem Reputation integration.",
                  "value": {
                    "cluster_id": "example",
                    "budget_types": {
                      "default": {
                        "budget": {
                          "golem_workers.budgets.AveragePerCpuUsageLinearModelBudget": {
                            "average_cpu_load": 1.0,
                            "average_duration_hours": 0.5,
                            "average_max_cost": 1.5
                          }
                        },
                        "scope": "cluster"
                      }
                    },
                    "network_types": {
                      "default": {
                        "ip": "192.168.0.0/16"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateClusterResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Unhandled server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPGenericError"
                }
              }
            }
          },
          "409": {
            "description": "Object already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPGenericError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cluster/{cluster_id}": {
      "get": {
        "tags": [
          "clusters"
        ],
        "summary": "Get Cluster",
        "description": "Get details for a specific cluster.",
        "operationId": "get_cluster_cluster__cluster_id__get",
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cluster Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetClusterResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Unhandled server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPGenericError"
                }
              }
            }
          },
          "404": {
            "description": "Object was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPGenericError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "clusters"
        ],
        "summary": "Delete Cluster",
        "description": "Delete a cluster.",
        "operationId": "delete_cluster_cluster__cluster_id__delete",
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cluster ID",
              "description": "cluster identifier given in create-cluster operation"
            },
            "description": "cluster identifier given in create-cluster operation",
            "example": "example"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteClusterResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Unhandled server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPGenericError"
                }
              }
            }
          },
          "404": {
            "description": "Object was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPGenericError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cluster/{cluster_id}/node": {
      "post": {
        "tags": [
          "nodes"
        ],
        "summary": "Create Node",
        "description": "Creates node. Apply logic from cluster configuration.",
        "operationId": "create_node_cluster__cluster_id__node_post",
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cluster ID",
              "description": "Cluster to which the new node will be attached"
            },
            "description": "Cluster to which the new node will be attached",
            "example": "example"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNodeRequest"
              },
              "examples": {
                "echo_test": {
                  "summary": "modelserve/echo-test:2",
                  "description": "This example shows how to run echo test with VPN",
                  "value": {
                    "cluster_id": "example",
                    "node_networks": {
                      "default": {}
                    },
                    "node_config": {
                      "market_config": {
                        "demand": {
                          "payloads": [
                            {
                              "golem_workers.payloads.ClusterNodePayload": {
                                "image_tag": "modelserve/echo-test:2"
                              }
                            }
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateNodeResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Unhandled server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPGenericError"
                }
              }
            }
          },
          "409": {
            "description": "Object already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPGenericError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cluster/{cluster_id}/node/{node_id}": {
      "get": {
        "tags": [
          "nodes"
        ],
        "summary": "Get Node",
        "description": "Get details for a specific node.",
        "operationId": "get_node_cluster__cluster_id__node__node_id__get",
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cluster Id"
            }
          },
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetNodeResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Unhandled server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPGenericError"
                }
              }
            }
          },
          "404": {
            "description": "Object was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPGenericError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "nodes"
        ],
        "summary": "Delete Node",
        "description": "Delete a node from a cluster.",
        "operationId": "delete_node_cluster__cluster_id__node__node_id__delete",
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cluster Id"
            }
          },
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteNodeResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Unhandled server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPGenericError"
                }
              }
            }
          },
          "404": {
            "description": "Object was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPGenericError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/ports/config": {
      "get": {
        "tags": [
          "ports"
        ],
        "summary": "Get Port Config",
        "description": "Returns the current port allocation service configuration.",
        "operationId": "get_port_config_ports_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortConfigResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Unhandled server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPGenericError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateClusterResponse": {
        "properties": {
          "cluster": {
            "$ref": "#/components/schemas/ClusterOut"
          }
        },
        "type": "object",
        "required": [
          "cluster"
        ],
        "title": "CreateClusterResponse"
      },
...
      "GetClusterResponse": {
        "properties": {
          "cluster": {
            "$ref": "#/components/schemas/ClusterOut"
          }
        },
        "type": "object",
        "required": [
          "cluster"
        ],
        "title": "GetClusterResponse"
      },
...
      "GetProposalsRequest": {
        "properties": {
          "market_config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Config"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "GetProposalsRequest"
      },
      "GetProposalsResponse": {
        "properties": {
          "proposals": {
            "items": {
              "$ref": "#/components/schemas/ProposalOut"
            },
            "type": "array",
            "title": "Proposals"
          }
        },
        "type": "object",
        "required": [
          "proposals"
        ],
        "title": "GetProposalsResponse"
      },
      "HTTPGenericError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "HTTPGenericError"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  },
...
}